!feat: pages for source/format view/create
Implemented the basic page structure to be used. Restructed/renamed
some routes existing routes, created more components and stories.

An unfortunate discovery is that Storybook doesn't work with the
Unplugin icons, so it's going to be necessary to replace them,
unless I can find a workaround. Discussion about it can be found in
[this GitHub issue](https://github.com/unplugin/unplugin-icons/issues/92).