feat: add various forms for doc and task entities
A fairly large commit due to various interconnected facets.

Create, Read and Delete now possible in the UI for every part of the
task creation process. Updates still stlll to come. Did a bit of
refactoring, but only where it would speed things up or ease them in
the short-term—there's more to do later.